refactor(user-skills): extract agent home owner
